+++ controller/controlplane/controlplane_controller_reconciler_utils_test.go
@@ -24,6 +24,7 @@ import (
 
 func TestEnsureClusterRole(t *testing.T) {
 	clusterRole, err := k8sresources.GenerateNewClusterRoleForControlPlane("test-controlplane", consts.DefaultControlPlaneImage, false)
+
 	assert.NoError(t, err)
 	clusterRole.Name = "test-clusterrole"
 	wrongClusterRole := clusterRole.DeepCopy()
@@ -43,6 +44,36 @@ func TestEnsureClusterRole(t *testing.T) {
 	wrongClusterRole2 := clusterRole.DeepCopy()
 	wrongClusterRole2.ObjectMeta.Labels["aaa"] = "bbb"
 
+	controlplane := operatorv1beta1.ControlPlane{
+		TypeMeta: metav1.TypeMeta{
+			APIVersion: "gateway-operator.konghq.com/v1beta1",
+			Kind:       "ControlPlane",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "test-controlplane",
+			Namespace: "test-namespace",
+			UID:       types.UID(uuid.NewString()),
+		},
+		Spec: operatorv1beta1.ControlPlaneSpec{
+			ControlPlaneOptions: operatorv1beta1.ControlPlaneOptions{
+				Deployment: operatorv1beta1.ControlPlaneDeploymentOptions{
+					PodTemplateSpec: &corev1.PodTemplateSpec{
+						Spec: corev1.PodSpec{
+							Containers: []corev1.Container{
+								{
+									Name:  consts.ControlPlaneControllerContainerName,
+									Image: consts.DefaultControlPlaneImage,
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+
+	k8sutils.SetOwnerForObjectThroughLabels(clusterRole, &controlplane)
+
 	testCases := []struct {
 		Name                string
 		controlplane        operatorv1beta1.ControlPlane
@@ -52,98 +83,20 @@ func TestEnsureClusterRole(t *testing.T) {
 		err                 error
 	}{
 		{
-			Name: "no existing clusterrole",
-			controlplane: operatorv1beta1.ControlPlane{
-				TypeMeta: metav1.TypeMeta{
-					APIVersion: "gateway-operator.konghq.com/v1beta1",
-					Kind:       "ControlPlane",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-controlplane",
-					Namespace: "test-namespace",
-					UID:       types.UID(uuid.NewString()),
-				},
-				Spec: operatorv1beta1.ControlPlaneSpec{
-					ControlPlaneOptions: operatorv1beta1.ControlPlaneOptions{
-						Deployment: operatorv1beta1.ControlPlaneDeploymentOptions{
-							PodTemplateSpec: &corev1.PodTemplateSpec{
-								Spec: corev1.PodSpec{
-									Containers: []corev1.Container{
-										{
-											Name:  consts.ControlPlaneControllerContainerName,
-											Image: consts.DefaultControlPlaneImage,
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
+			Name:                "no existing clusterrole",
+			controlplane:        controlplane,
 			createdorUpdated:    true,
 			expectedClusterRole: *clusterRole,
 		},
 		{
-			Name: "up to date clusterrole",
-			controlplane: operatorv1beta1.ControlPlane{
-				TypeMeta: metav1.TypeMeta{
-					APIVersion: "gateway-operator.konghq.com/v1beta1",
-					Kind:       "ControlPlane",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-controlplane",
-					Namespace: "test-namespace",
-					UID:       types.UID(uuid.NewString()),
-				},
-				Spec: operatorv1beta1.ControlPlaneSpec{
-					ControlPlaneOptions: operatorv1beta1.ControlPlaneOptions{
-						Deployment: operatorv1beta1.ControlPlaneDeploymentOptions{
-							PodTemplateSpec: &corev1.PodTemplateSpec{
-								Spec: corev1.PodSpec{
-									Containers: []corev1.Container{
-										{
-											Name:  consts.ControlPlaneControllerContainerName,
-											Image: consts.DefaultControlPlaneImage,
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
+			Name:                "up to date clusterrole",
+			controlplane:        controlplane,
 			existingClusterRole: clusterRole,
 			expectedClusterRole: *clusterRole,
 		},
 		{
-			Name: "out of date clusterrole, object meta",
-			controlplane: operatorv1beta1.ControlPlane{
-				TypeMeta: metav1.TypeMeta{
-					APIVersion: "gateway-operator.konghq.com/v1beta1",
-					Kind:       "ControlPlane",
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-controlplane",
-					Namespace: "test-namespace",
-					UID:       types.UID(uuid.NewString()),
-				},
-				Spec: operatorv1beta1.ControlPlaneSpec{
-					ControlPlaneOptions: operatorv1beta1.ControlPlaneOptions{
-						Deployment: operatorv1beta1.ControlPlaneDeploymentOptions{
-							PodTemplateSpec: &corev1.PodTemplateSpec{
-								Spec: corev1.PodSpec{
-									Containers: []corev1.Container{
-										{
-											Name:  consts.ControlPlaneControllerContainerName,
-											Image: consts.DefaultControlPlaneImage,
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
+			Name:                "out of date clusterrole, object meta",
+			controlplane:        controlplane,
 			existingClusterRole: wrongClusterRole2,
 			createdorUpdated:    true,
 			expectedClusterRole: *clusterRole,
@@ -158,7 +111,7 @@ func TestEnsureClusterRole(t *testing.T) {
 		}
 
 		if tc.existingClusterRole != nil {
-			k8sutils.SetOwnerForObject(tc.existingClusterRole, &tc.controlplane)
+			k8sutils.SetOwnerForObjectThroughLabels(tc.existingClusterRole, &tc.controlplane)
 			ObjectsToAdd = append(ObjectsToAdd, tc.existingClusterRole)
 		}
 
@@ -296,7 +249,7 @@ func TestEnsureClusterRoleBinding(t *testing.T) {
 			controlPlane,
 		}
 		for _, crb := range tc.existingCRBs {
-			k8sutils.SetOwnerForObject(crb, controlPlane)
+			k8sutils.SetOwnerForObjectThroughLabels(crb, controlPlane)
 			objectsToAdd = append(objectsToAdd, crb)
 		}